Rename input folder and move CAS to separate component]
