+++ frontend/app/(tabs)/addChemical.tsx
@@ -1,68 +1,17 @@
 import CustomButton from '@/components/CustomButton';
-import CustomTextInput from '@/components/CustomTextInput';
+import HeaderTextInput from '@/components/inputFields/HeaderTextInput';
 import Header from '@/components/Header';
-import CustomTextBox from '@/components/inputField/CustomTextBox';
-import CustomTextHeader from '@/components/inputField/CustomTextHeader';
-import React, { RefObject, useRef, useState } from 'react';
+import CustomTextHeader from '@/components/inputFields/CustomTextHeader';
+import React, { useState } from 'react';
 import { View, StyleSheet, ScrollView, Text, TextInput } from 'react-native';
+import CasTextBoxes from '@/components/inputFields/CasTextBoxes';
 
 export default function ViewChemicals() {
   const [name, setName] = useState<string>('')
-
-  // stuff for shifting focus and keeping track of CAS number inputs
+  const [room, setRoom] = useState<string>('')
+  const [shelf, setShelf] = useState<string>('')
+  const [cabinet, setCabinet] = useState<string>('')
   const [casParts, setCasParts] = useState<string[]>(['', '', ''])
-  const casRefs: RefObject<TextInput>[] = [
-    useRef<TextInput>(null),
-    useRef<TextInput>(null),
-    useRef<TextInput>(null)]
-
-  const Divider = () => {
-    return (
-      <View style={styles.line} />
-    )
-  }
-
-  // for when we need to shift focus to another section for CAS number
-  const onCasChange = (value: string, index: number, maxLength: number) => {
-    const newCasParts = [...casParts]
-    newCasParts[index] = value
-    setCasParts(newCasParts)
-
-    if (value.length === maxLength && index < casParts.length - 1) {
-      casRefs[index + 1].current?.focus()
-    }
-  }
-
-  // handle backspace, shift focus to previous textbox
-  const onCasKeyPress = (e: any, index: number) => {
-    if (e.nativeEvent.key == 'Backspace' && casParts[index] == '' && index > 0) {
-      casRefs[index - 1].current?.focus()
-    }
-  }
-
-  // handle next button, shift to next textbox, or if last textbox do nothing
-  const onCasSubmit = (index : number) => {
-    if (index < casParts.length - 1) {
-      casRefs[index + 1].current?.focus()
-    } else {
-      console.log(casParts)
-    }
-  }
-
-  // helper function for the CAS number text box
-  const CasTextBox = (index: number, width: number, maxLength: number, isNext: boolean) => (
-    <CustomTextBox
-      keyboardType='numeric'
-      onChangeText={(value: string) => onCasChange(value, index, maxLength)}
-      ref={casRefs[index]}
-      width={width}
-      isCenter={true}
-      maxLength={maxLength}
-      onKeyPress={(e: any) => {onCasKeyPress(e, index)}}
-      returnKeyType={isNext ? 'next' : 'done'}
-      onSubmitEditing={() => {onCasSubmit(index)}}
-    />
-  )
 
   return (
     <View style={styles.container}>
@@ -71,24 +20,24 @@ export default function ViewChemicals() {
         <View style={styles.innerContainer}>
 
           {/* Name */}
-          <CustomTextInput
+          <HeaderTextInput
             headerText='Name'
             onChangeText={(value: string) => { setName(value) }}
           />
 
-          {/* CAS Number */}
+          {/* CAS Number */ }
           <CustomTextHeader headerText='CAS Number' />
-          <View style={styles.cas}>
-            {CasTextBox(0, 135, 7, true)}
-            <Divider />
-            {CasTextBox(1, 75, 2, true)}
-            <Divider />
-            {CasTextBox(2, 40, 1, false)}
+          <CasTextBoxes casParts={casParts} setCasParts={setCasParts} />
 
+          {/* Room, cabinet, shelf number */}
+          <View style={styles.row}>
+            <HeaderTextInput headerText={'Room'} onChangeText={(value : string) => setRoom(value)} inputWidth={111} />
+            <HeaderTextInput headerText={'Shelf'} onChangeText={(value : string) => setCabinet(value)} inputWidth={88} />
+            <HeaderTextInput headerText={'Cabinet'} onChangeText={(value : string) => setShelf(value)} inputWidth={88} />
           </View>
 
           {/* Button to test text input */}
-          <CustomButton title={'Test'} onPress={() => { console.log(JSON.stringify(casParts)) }} width={75} />
+          <CustomButton title={'Test'} onPress={() => { console.log(JSON.stringify(casParts)) }} width={80} />
         </View>
       </ScrollView>
     </View>
@@ -104,22 +53,15 @@ const styles = StyleSheet.create({
   },
   innerContainer: {
     marginTop: 150,
-    marginHorizontal: '8%'
+    marginHorizontal: '8%',
   },
   scroll: {
     width: '100%',
   },
-  cas: {
+  row: {
     display: 'flex',
     flexDirection: 'row',
     width: '100%',
     justifyContent: 'space-between',
-  },
-  line: {
-    height: 2,
-    backgroundColor: '#000000',
-    width: 10,
-    alignSelf: 'center',
-    borderRadius: 2,
   },
 });
